+++ src/iife.js
@@ -9,6 +9,7 @@ import './elementals/checkbox-group/index.js';
 import './elementals/combobox/index.js';
 import './elementals/copy/index.js';
 import './elementals/disclosure/index.js';
+import './elementals/field/index.js';
 import './elementals/marquee/index.js';
 import './elementals/menu/index.js';
 import './elementals/modal/index.js';